Split plugin and runtime distribution CLI responsibilities
Amp-Thread-ID: https://ampcode.com/threads/T-01a012e0-f664-73c8-b226-043c8eb091a2
