+++ internal/cli/cli.go
@@ -2,19 +2,14 @@
 package cli
 
 import (
-	"archive/tar"
 	"bytes"
-	"compress/gzip"
 	"context"
 	"crypto/sha256"
-	"debug/elf"
-	"debug/macho"
 	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
-	"net/http"
 	"os"
 	"os/exec"
 	"os/signal"
@@ -26,7 +21,6 @@ import (
 	"strings"
 	"sync"
 	"syscall"
-	"time"
 
 	actionintegration "github.com/buildkite/buildkite-gha/internal/action/integration"
 	actionsource "github.com/buildkite/buildkite-gha/internal/action/source"
@@ -41,31 +35,21 @@ import (
 )
 
 const (
-	legacyRuntimeQueue                = "hosted"
-	pluginConfigurationEnvironment    = "BUILDKITE_PLUGIN_CONFIGURATION"
-	pluginDevDarwinRuntimeEnvironment = "BUILDKITE_GHA_PLUGIN_DEV_DARWIN_RUNTIME"
-	pluginDevLinuxRuntimeEnvironment  = "BUILDKITE_GHA_PLUGIN_DEV_LINUX_RUNTIME"
-	legacyTargetQueueEnvironment      = "BUILDKITE_GHA_TARGET_QUEUE"
-	legacyRuntimeImageEnvironment     = "BUILDKITE_GHA_RUNTIME_IMAGE"
-	hostedProfile                     = "hosted"
-	legacyHostedTokenlessProfile      = "hosted-tokenless"
-	runtimeDistributionLimit          = 256 << 20
-	pluginChecksumLimit               = 4 << 20
-	pluginArchiveLimit                = 256 << 20
-	maxWebhookMetadataBytes           = 25 << 20
-	workflowCheckSummaryLimit         = 65535
-	workflowCheckSummaryNotice        = "\n\n_Additional diagnostics omitted at the provider check summary size limit._\n"
-	defaultNobleRunnerImage           = "buildkite.namespace-images.com/agent-base@sha256:d8fa20ce298d9ee8673dc6e9ecddabb23ebbff3bcaf828d2011714a960cc5853"
-	defaultJammyRunnerImage           = "buildkite.namespace-images.com/agent-base@sha256:3489d7d999869a11a243bdd9f91f318784cb74f0b57646c57424d10008b1e21d"
-	defaultMacOSRunnerQueue           = "macos-medium"
+	legacyRuntimeQueue            = "hosted"
+	legacyTargetQueueEnvironment  = "BUILDKITE_GHA_TARGET_QUEUE"
+	legacyRuntimeImageEnvironment = "BUILDKITE_GHA_RUNTIME_IMAGE"
+	hostedProfile                 = "hosted"
+	legacyHostedTokenlessProfile  = "hosted-tokenless"
+	maxWebhookMetadataBytes       = 25 << 20
+	workflowCheckSummaryLimit     = 65535
+	workflowCheckSummaryNotice    = "\n\n_Additional diagnostics omitted at the provider check summary size limit._\n"
+	defaultNobleRunnerImage       = "buildkite.namespace-images.com/agent-base@sha256:d8fa20ce298d9ee8673dc6e9ecddabb23ebbff3bcaf828d2011714a960cc5853"
+	defaultJammyRunnerImage       = "buildkite.namespace-images.com/agent-base@sha256:3489d7d999869a11a243bdd9f91f318784cb74f0b57646c57424d10008b1e21d"
+	defaultMacOSRunnerQueue       = "macos-medium"
 )
 
 var runnerQueuePattern = regexp.MustCompile(`^[A-Za-z0-9_-]{1,255}$`)
 var runnerImagePattern = regexp.MustCompile(`^[a-z0-9]+(?:[._-][a-z0-9]+)*(?:/[a-z0-9]+(?:[._-][a-z0-9]+)*)+@sha256:[0-9a-f]{64}$`)
-var stableVersionPattern = regexp.MustCompile(`^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)$`)
-
-var pluginReleaseBaseURL = "https://github.com/buildkite/buildkite-gha/releases/download"
-var pluginHTTPClient = securePluginHTTPClient()
 
 // Run executes the command and returns its process exit code.
 func Run(args []string, stdout, stderr io.Writer, version string) int {
@@ -128,250 +112,6 @@ func commandVersion(clientVersion string) string {
 		return "dev"
 	}
 	return clientVersion
-}
-
-func plugin(args []string, stdout, stderr io.Writer, version, clientVersion string, runner transport.Runner) int {
-	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
-	defer stop()
-	return pluginContext(ctx, args, stdout, stderr, version, clientVersion, runner)
-}
-
-func pluginContext(ctx context.Context, args []string, stdout, stderr io.Writer, version, clientVersion string, runner transport.Runner) (code int) {
-	started := time.Now()
-	details := &commandTelemetryDetails{}
-	defer func() {
-		outcome := telemetryOutcome(code, "", ctx.Err())
-		emitCommandTelemetry(telemetry.CommandPluginImport, outcome, clientVersion, time.Since(started), details.forOutcome(outcome))
-	}()
-	if len(args) != 0 {
-		return usageError(stderr, "plugin does not accept arguments")
-	}
-	importerPlatform, err := importerPlatform(runtime.GOOS, runtime.GOARCH)
-	if err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: plugin: %v\n", err)
-		return 1
-	}
-	configuration, err := parsePluginConfiguration(os.Getenv(pluginConfigurationEnvironment))
-	if err != nil {
-		return usageError(stderr, "plugin: %v", err)
-	}
-	if err := normalizePluginCommit(ctx, os.Getenv, os.Setenv, runner); err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: plugin: %v\n", err)
-		return 1
-	}
-	return uploadParsedContext(ctx, parsedUploadArgs{
-		workflowOperands:       configuration.Workflows,
-		explicitWorkflowPaths:  true,
-		runnerTargets:          configuration.runnerTargets,
-		oidc:                   configuration.OIDC,
-		experimentalRunnerUser: configuration.ExperimentalRunnerUser,
-		pluginAcquisition:      &pluginRuntimeAcquisition{version: version},
-		importerPlatform:       importerPlatform,
-		telemetry:              details,
-	}, stdout, stderr, version, transport.Agent{Runner: runner})
-}
-
-func importerPlatform(goos, goarch string) (compiler.Platform, error) {
-	switch {
-	case goos == "linux" && goarch == "amd64":
-		return compiler.PlatformLinuxAMD64, nil
-	case goos == "darwin" && goarch == "arm64":
-		return compiler.PlatformDarwinARM64, nil
-	default:
-		return compiler.Platform{}, fmt.Errorf("importer requires linux/amd64 or darwin/arm64, running on %s/%s", goos, goarch)
-	}
-}
-
-type pluginConfiguration struct {
-	Workflows              []string
-	ExperimentalRunnerUser bool
-	OIDC                   *plan.OIDCConfiguration
-	runnerTargets          map[string]compiler.RunnerTarget
-}
-
-func parsePluginConfiguration(source string) (pluginConfiguration, error) {
-	if strings.TrimSpace(source) == "" {
-		return pluginConfiguration{}, fmt.Errorf("%s is required", pluginConfigurationEnvironment)
-	}
-	decoder := json.NewDecoder(strings.NewReader(source))
-	value, err := decodeJSONValue(decoder)
-	if err != nil {
-		return pluginConfiguration{}, fmt.Errorf("decode %s: %w", pluginConfigurationEnvironment, err)
-	}
-	if _, err := decoder.Token(); err != io.EOF {
-		if err == nil {
-			return pluginConfiguration{}, fmt.Errorf("decode %s: multiple JSON values", pluginConfigurationEnvironment)
-		}
-		return pluginConfiguration{}, fmt.Errorf("decode %s: %w", pluginConfigurationEnvironment, err)
-	}
-	encoded, ok := value.(map[string]any)
-	if !ok {
-		return pluginConfiguration{}, fmt.Errorf("%s must be a JSON object", pluginConfigurationEnvironment)
-	}
-	for key := range encoded {
-		switch key {
-		case "workflow", "workflows", "runners", "oidc", "version", "source-ref", "minimum-release-age", "experimental-runner-user":
-		default:
-			return pluginConfiguration{}, fmt.Errorf("%s contains unknown field %q", pluginConfigurationEnvironment, key)
-		}
-	}
-	experimentalRunnerUser := true
-	if value, configured := encoded["experimental-runner-user"]; configured {
-		var ok bool
-		experimentalRunnerUser, ok = value.(bool)
-		if !ok {
-			return pluginConfiguration{}, fmt.Errorf("%s experimental-runner-user must be a boolean", pluginConfigurationEnvironment)
-		}
-	}
-	workflowValue, hasWorkflow := encoded["workflow"]
-	workflowsValue, hasWorkflows := encoded["workflows"]
-	if hasWorkflow && hasWorkflows {
-		return pluginConfiguration{}, fmt.Errorf("%s workflow and workflows are mutually exclusive", pluginConfigurationEnvironment)
-	}
-	var workflows []string
-	if hasWorkflow {
-		workflow, ok := workflowValue.(string)
-		if !ok || strings.TrimSpace(workflow) == "" {
-			return pluginConfiguration{}, fmt.Errorf("%s workflow must be a non-empty string", pluginConfigurationEnvironment)
-		}
-		workflows = []string{workflow}
-	} else {
-		values, ok := workflowsValue.([]any)
-		if ok && len(values) != 0 {
-			workflows = make([]string, len(values))
-			for index, entry := range values {
-				workflow, ok := entry.(string)
-				if !ok || strings.TrimSpace(workflow) == "" {
-					return pluginConfiguration{}, fmt.Errorf("%s workflows entry %d must be a non-empty string", pluginConfigurationEnvironment, index)
-				}
-				workflows[index] = workflow
-			}
-		}
-		if len(workflows) == 0 {
-			return pluginConfiguration{}, fmt.Errorf("%s workflow or workflows is required; workflows must be a non-empty array of non-empty strings", pluginConfigurationEnvironment)
-		}
-	}
-	targets := make(map[string]compiler.RunnerTarget)
-	if runnersValue, configured := encoded["runners"]; configured {
-		runners, ok := runnersValue.([]any)
-		if !ok || len(runners) == 0 {
-			return pluginConfiguration{}, fmt.Errorf("%s runners must be a non-empty array when configured", pluginConfigurationEnvironment)
-		}
-		for index, runnerValue := range runners {
-			runner, ok := runnerValue.(map[string]any)
-			if !ok {
-				return pluginConfiguration{}, fmt.Errorf("runner %d must be a JSON object", index)
-			}
-			for key := range runner {
-				switch key {
-				case "runs-on", "queue", "image":
-				default:
-					return pluginConfiguration{}, fmt.Errorf("runner %d contains unknown field %q", index, key)
-				}
-			}
-			runsOn, runsOnOK := runner["runs-on"].(string)
-			queue, queueOK := runner["queue"].(string)
-			if !runsOnOK {
-				return pluginConfiguration{}, fmt.Errorf("runner %d runs-on must be a string", index)
-			}
-			if !queueOK {
-				return pluginConfiguration{}, fmt.Errorf("runner %d queue must be a string", index)
-			}
-			image := ""
-			if imageValue, configured := runner["image"]; configured {
-				var imageOK bool
-				image, imageOK = imageValue.(string)
-				if !imageOK || image == "" {
-					return pluginConfiguration{}, fmt.Errorf("runner %d image must be an immutable registry sha256 reference", index)
-				}
-			}
-			label, target, err := configuredRunnerTarget(runsOn, queue, image)
-			if err != nil {
-				return pluginConfiguration{}, fmt.Errorf("runner %d: %w", index, err)
-			}
-			if _, duplicate := targets[label]; duplicate {
-				return pluginConfiguration{}, fmt.Errorf("runner label %q may only be configured once", label)
-			}
-			targets[label] = target
-		}
-	}
-	var oidc *plan.OIDCConfiguration
-	if value, configured := encoded["oidc"]; configured {
-		oidc, err = parsePluginOIDCConfiguration(value)
-		if err != nil {
-			return pluginConfiguration{}, err
-		}
-	}
-	return pluginConfiguration{Workflows: workflows, ExperimentalRunnerUser: experimentalRunnerUser, OIDC: oidc, runnerTargets: targets}, nil
-}
-
-func parsePluginOIDCConfiguration(value any) (*plan.OIDCConfiguration, error) {
-	oidc, ok := value.(map[string]any)
-	if !ok {
-		return nil, fmt.Errorf("%s oidc must be a JSON object", pluginConfigurationEnvironment)
-	}
-	for key := range oidc {
-		switch key {
-		case "claims", "aws-session-tags", "subject-claim":
-		default:
-			return nil, fmt.Errorf("%s oidc contains unknown field %q", pluginConfigurationEnvironment, key)
-		}
-	}
-	configuration := &plan.OIDCConfiguration{}
-	var err error
-	if claims, configured := oidc["claims"]; configured {
-		configuration.Claims, err = pluginNonEmptyStringList(claims, "oidc claims")
-		if err != nil {
-			return nil, err
-		}
-	}
-	if tags, configured := oidc["aws-session-tags"]; configured {
-		configuration.AWSSessionTags, err = pluginNonEmptyStringList(tags, "oidc aws-session-tags")
-		if err != nil {
-			return nil, err
-		}
-	}
-	if subject, configured := oidc["subject-claim"]; configured {
-		configuration.SubjectClaim, ok = subject.(string)
-		if !ok || strings.TrimSpace(configuration.SubjectClaim) == "" {
-			return nil, fmt.Errorf("%s oidc subject-claim must be a non-empty string", pluginConfigurationEnvironment)
-		}
-	}
-	return configuration, nil
-}
-
-func pluginNonEmptyStringList(value any, field string) ([]string, error) {
-	values, ok := value.([]any)
-	if !ok || len(values) == 0 {
-		return nil, fmt.Errorf("%s %s must be a non-empty array of non-empty strings", pluginConfigurationEnvironment, field)
-	}
-	result := make([]string, len(values))
-	for index, value := range values {
-		entry, ok := value.(string)
-		if !ok || strings.TrimSpace(entry) == "" {
-			return nil, fmt.Errorf("%s %s entry %d must be a non-empty string", pluginConfigurationEnvironment, field, index)
-		}
-		result[index] = entry
-	}
-	return result, nil
-}
-
-func normalizePluginCommit(ctx context.Context, getenv func(string) string, setenv func(string, string) error, runner transport.Runner) error {
-	if validBuildkiteCommit(getenv("BUILDKITE_COMMIT")) {
-		return nil
-	}
-	output, err := runner.Run(ctx, "", "git", []string{"rev-parse", "HEAD"}, nil)
-	if err != nil {
-		return fmt.Errorf("resolve BUILDKITE_COMMIT from checked-out HEAD: %w", err)
-	}
-	commit := strings.TrimSpace(string(output))
-	if !validBuildkiteCommit(commit) {
-		return fmt.Errorf("resolve BUILDKITE_COMMIT from checked-out HEAD: git returned an invalid commit")
-	}
-	if err := setenv("BUILDKITE_COMMIT", commit); err != nil {
-		return fmt.Errorf("set resolved BUILDKITE_COMMIT: %w", err)
-	}
-	return nil
 }
 
 func validate(args []string, stdout, stderr io.Writer, version string, agent transport.Agent) int {
@@ -2250,347 +1990,6 @@ func supportedRunnerTarget(label string) (string, compiler.Platform, error) {
 	}
 }
 
-type runtimeDistribution struct {
-	contents []byte
-	digest   string
-}
-
-type pluginRuntimeAcquisition struct {
-	version string
-}
-
-const (
-	pluginLinuxAsset  = "buildkite-gha_Linux_x86_64.tar.gz"
-	pluginDarwinAsset = "buildkite-gha_Darwin_arm64.tar.gz"
-)
-
-func securePluginHTTPClient() *http.Client {
-	return &http.Client{
-		Timeout: 2 * time.Minute,
-		CheckRedirect: func(request *http.Request, via []*http.Request) error {
-			if request.URL.Scheme != "https" {
-				return errors.New("release download redirected away from HTTPS")
-			}
-			if len(via) >= 10 {
-				return errors.New("too many release download redirects")
-			}
-			return nil
-		},
-	}
-}
-
-func (a *pluginRuntimeAcquisition) acquire(ctx context.Context, required map[compiler.Platform]bool, hostPlatform compiler.Platform, host runtimeDistribution) (map[compiler.Platform]runtimeDistribution, error) {
-	distributions := make(map[compiler.Platform]runtimeDistribution, len(required))
-	devPaths := map[compiler.Platform]string{
-		compiler.PlatformLinuxAMD64:  os.Getenv(pluginDevLinuxRuntimeEnvironment),
-		compiler.PlatformDarwinARM64: os.Getenv(pluginDevDarwinRuntimeEnvironment),
-	}
-	if a.version != "dev" {
-		for platform, path := range devPaths {
-			if path != "" {
-				return nil, fmt.Errorf("%s is test/dev-only and is rejected by release builds", pluginDevRuntimeEnvironment(platform))
-			}
-		}
-	}
-	for _, platform := range []compiler.Platform{compiler.PlatformLinuxAMD64, compiler.PlatformDarwinARM64} {
-		if !required[platform] {
-			continue
-		}
-		if platform == hostPlatform {
-			distributions[platform] = host
-			continue
-		}
-		if a.version == "dev" {
-			path := devPaths[platform]
-			if path == "" {
-				return nil, fmt.Errorf("%s is required for a dev build using %s", pluginDevRuntimeEnvironment(platform), platform)
-			}
-			loaded, err := loadRuntimeDistributions(map[compiler.Platform]string{platform: path})
-			if err != nil {
-				return nil, err
-			}
-			distributions[platform] = loaded[platform]
-			continue
-		}
-		if !stableVersionPattern.MatchString(a.version) {
-			return nil, fmt.Errorf("running version %q is not a stable semantic version", a.version)
-		}
-		distribution, err := acquirePluginRuntime(ctx, a.version, platform, pluginHTTPClient, pluginReleaseBaseURL, pluginArchiveCachePath(a.version, pluginRuntimeAsset(platform)))
-		if err != nil {
-			return nil, err
-		}
-		distributions[platform] = distribution
-	}
-	return distributions, nil
-}
-
-func pluginDevRuntimeEnvironment(platform compiler.Platform) string {
-	if platform == compiler.PlatformDarwinARM64 {
-		return pluginDevDarwinRuntimeEnvironment
-	}
-	return pluginDevLinuxRuntimeEnvironment
-}
-
-func pluginRuntimeAsset(platform compiler.Platform) string {
-	if platform == compiler.PlatformDarwinARM64 {
-		return pluginDarwinAsset
-	}
-	return pluginLinuxAsset
-}
-
-func acquirePluginRuntime(ctx context.Context, version string, platform compiler.Platform, client *http.Client, baseURL, cachePath string) (runtimeDistribution, error) {
-	asset := pluginRuntimeAsset(platform)
-	checksums, err := downloadPluginReleaseFile(ctx, client, baseURL+"/v"+version+"/checksums.txt", pluginChecksumLimit)
-	if err != nil {
-		return runtimeDistribution{}, fmt.Errorf("download release checksums: %w", err)
-	}
-	expected, err := pluginAssetChecksum(checksums, asset)
-	if err != nil {
-		return runtimeDistribution{}, err
-	}
-	archive := readVerifiedPluginArchiveCache(cachePath, expected)
-	if archive == nil {
-		archive, err = downloadPluginReleaseFile(ctx, client, baseURL+"/v"+version+"/"+asset, pluginArchiveLimit)
-		if err != nil {
-			return runtimeDistribution{}, fmt.Errorf("download %s runtime: %w", platform, err)
-		}
-		if sha256.Sum256(archive) != expected {
-			return runtimeDistribution{}, fmt.Errorf("%s archive checksum verification failed", platform)
-		}
-		writePluginArchiveCache(cachePath, archive)
-	}
-	contents, err := extractPluginRuntime(archive, platform)
-	if err != nil {
-		return runtimeDistribution{}, err
-	}
-	if err := validateRuntimeDistributionBinary(platform, contents); err != nil {
-		return runtimeDistribution{}, fmt.Errorf("validate %s runtime: %w", platform, err)
-	}
-	digest := sha256.Sum256(contents)
-	return runtimeDistribution{contents: contents, digest: fmt.Sprintf("sha256:%x", digest)}, nil
-}
-
-func downloadPluginReleaseFile(ctx context.Context, client *http.Client, source string, limit int64) ([]byte, error) {
-	request, err := http.NewRequestWithContext(ctx, http.MethodGet, source, nil)
-	if err != nil || request.URL.Scheme != "https" {
-		return nil, fmt.Errorf("release URL must use HTTPS")
-	}
-	response, err := client.Do(request)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = response.Body.Close() }()
-	if response.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("unexpected HTTP status %s", response.Status)
-	}
-	contents, err := io.ReadAll(io.LimitReader(response.Body, limit+1))
-	if err != nil {
-		return nil, err
-	}
-	if int64(len(contents)) > limit {
-		return nil, fmt.Errorf("exceeds %d-byte limit", limit)
-	}
-	return contents, nil
-}
-
-func pluginAssetChecksum(contents []byte, asset string) ([sha256.Size]byte, error) {
-	var result [sha256.Size]byte
-	matches := 0
-	for _, line := range strings.Split(string(contents), "\n") {
-		fields := strings.Fields(line)
-		if len(fields) != 2 || strings.TrimPrefix(fields[1], "*") != asset {
-			continue
-		}
-		decoded, err := hex.DecodeString(fields[0])
-		if err != nil || len(decoded) != sha256.Size {
-			return result, fmt.Errorf("invalid checksum entry for %s", asset)
-		}
-		copy(result[:], decoded)
-		matches++
-	}
-	if matches != 1 {
-		return result, fmt.Errorf("checksums must contain exactly one valid entry for %s", asset)
-	}
-	return result, nil
-}
-
-func extractPluginRuntime(archive []byte, platform compiler.Platform) ([]byte, error) {
-	gzipReader, err := gzip.NewReader(bytes.NewReader(archive))
-	if err != nil {
-		return nil, fmt.Errorf("open %s archive: %w", platform, err)
-	}
-	defer func() { _ = gzipReader.Close() }()
-	tarReader := tar.NewReader(gzipReader)
-	seenBinary, seenLicense := false, false
-	var executable []byte
-	for {
-		header, err := tarReader.Next()
-		if errors.Is(err, io.EOF) {
-			break
-		}
-		if err != nil {
-			return nil, fmt.Errorf("read %s archive: %w", platform, err)
-		}
-		if header.Typeflag != tar.TypeReg || header.Size < 0 || header.Size > runtimeDistributionLimit {
-			return nil, fmt.Errorf("%s archive contains an unsafe member %q", platform, header.Name)
-		}
-		switch header.Name {
-		case "buildkite-gha":
-			if seenBinary || header.Mode&0o111 == 0 {
-				return nil, fmt.Errorf("%s archive has an invalid executable member", platform)
-			}
-			seenBinary = true
-			executable, err = io.ReadAll(io.LimitReader(tarReader, runtimeDistributionLimit+1))
-			if err != nil || int64(len(executable)) != header.Size {
-				return nil, fmt.Errorf("read %s executable", platform)
-			}
-		case "LICENSE":
-			if seenLicense {
-				return nil, fmt.Errorf("%s archive contains duplicate LICENSE", platform)
-			}
-			seenLicense = true
-		default:
-			return nil, fmt.Errorf("%s archive contains unexpected member %q", platform, header.Name)
-		}
-	}
-	if !seenBinary || !seenLicense {
-		return nil, fmt.Errorf("%s archive must contain exactly buildkite-gha and LICENSE", platform)
-	}
-	return executable, nil
-}
-
-func pluginArchiveCachePath(version, asset string) string {
-	root := os.Getenv("MISE_DATA_DIR")
-	if !filepath.IsAbs(root) {
-		return ""
-	}
-	return filepath.Join(root, "buildkite-gha", "releases", version, asset)
-}
-
-func readVerifiedPluginArchiveCache(path string, expected [sha256.Size]byte) []byte {
-	if path == "" {
-		return nil
-	}
-	info, err := os.Lstat(path)
-	if err != nil || !info.Mode().IsRegular() || info.Size() <= 0 || info.Size() > pluginArchiveLimit {
-		return nil
-	}
-	contents, err := os.ReadFile(path)
-	if err != nil || sha256.Sum256(contents) != expected {
-		return nil
-	}
-	return contents
-}
-
-func writePluginArchiveCache(path string, contents []byte) {
-	if path == "" {
-		return
-	}
-	directory := filepath.Dir(path)
-	if err := os.MkdirAll(directory, 0o700); err != nil {
-		return
-	}
-	realDirectory, err := filepath.EvalSymlinks(directory)
-	if err != nil || realDirectory != directory {
-		return
-	}
-	info, err := os.Lstat(directory)
-	if err != nil || !info.IsDir() || info.Mode()&os.ModeSymlink != 0 {
-		return
-	}
-	temporary, err := os.CreateTemp(directory, ".archive-")
-	if err != nil {
-		return
-	}
-	name := temporary.Name()
-	defer func() { _ = os.Remove(name) }()
-	if _, err = temporary.Write(contents); err != nil || temporary.Sync() != nil || temporary.Close() != nil {
-		_ = temporary.Close()
-		return
-	}
-	_ = os.Rename(name, path)
-}
-
-func loadRuntimeDistributions(paths map[compiler.Platform]string) (map[compiler.Platform]runtimeDistribution, error) {
-	distributions := make(map[compiler.Platform]runtimeDistribution, len(paths))
-	for platform, path := range paths {
-		if !filepath.IsAbs(path) {
-			return nil, fmt.Errorf("runtime distribution for %s must use an absolute path", platform)
-		}
-		pathInfo, err := os.Lstat(path)
-		if err != nil {
-			return nil, fmt.Errorf("inspect runtime distribution for %s: %w", platform, err)
-		}
-		if pathInfo.Mode()&os.ModeSymlink != 0 {
-			return nil, fmt.Errorf("runtime distribution for %s is not a non-symlink executable regular file", platform)
-		}
-		file, err := os.Open(path)
-		if err != nil {
-			return nil, fmt.Errorf("open runtime distribution for %s: %w", platform, err)
-		}
-		info, err := file.Stat()
-		if err != nil {
-			_ = file.Close()
-			return nil, fmt.Errorf("inspect opened runtime distribution for %s: %w", platform, err)
-		}
-		if !os.SameFile(pathInfo, info) {
-			_ = file.Close()
-			return nil, fmt.Errorf("runtime distribution for %s changed while being opened", platform)
-		}
-		if !info.Mode().IsRegular() || info.Mode().Perm()&0o111 == 0 {
-			_ = file.Close()
-			return nil, fmt.Errorf("runtime distribution for %s is not a non-symlink executable regular file", platform)
-		}
-		if info.Size() <= 0 || info.Size() > runtimeDistributionLimit {
-			_ = file.Close()
-			return nil, fmt.Errorf("runtime distribution for %s must be between 1 and %d bytes", platform, runtimeDistributionLimit)
-		}
-		contents, err := io.ReadAll(io.LimitReader(file, runtimeDistributionLimit+1))
-		closeErr := file.Close()
-		if err != nil {
-			return nil, fmt.Errorf("read runtime distribution for %s: %w", platform, err)
-		}
-		if closeErr != nil {
-			return nil, fmt.Errorf("close runtime distribution for %s: %w", platform, closeErr)
-		}
-		if int64(len(contents)) != info.Size() {
-			return nil, fmt.Errorf("runtime distribution for %s changed while being read", platform)
-		}
-		if err := validateRuntimeDistributionBinary(platform, contents); err != nil {
-			return nil, fmt.Errorf("validate runtime distribution for %s: %w", platform, err)
-		}
-		sum := sha256.Sum256(contents)
-		distributions[platform] = runtimeDistribution{contents: contents, digest: fmt.Sprintf("sha256:%x", sum)}
-	}
-	return distributions, nil
-}
-
-func validateRuntimeDistributionBinary(platform compiler.Platform, contents []byte) error {
-	switch platform {
-	case compiler.PlatformLinuxAMD64:
-		binary, err := elf.NewFile(bytes.NewReader(contents))
-		if err != nil {
-			return fmt.Errorf("open ELF executable: %w", err)
-		}
-		defer func() { _ = binary.Close() }()
-		if binary.Class != elf.ELFCLASS64 || binary.Machine != elf.EM_X86_64 || binary.Type != elf.ET_EXEC {
-			return fmt.Errorf("want a thin 64-bit linux/amd64 executable")
-		}
-	case compiler.PlatformDarwinARM64:
-		binary, err := macho.NewFile(bytes.NewReader(contents))
-		if err != nil {
-			return fmt.Errorf("open Mach-O executable: %w", err)
-		}
-		defer func() { _ = binary.Close() }()
-		if binary.Cpu != macho.CpuArm64 || binary.Type != macho.TypeExec {
-			return fmt.Errorf("want a thin darwin/arm64 executable")
-		}
-	default:
-		return fmt.Errorf("unsupported platform")
-	}
-	return nil
-}
-
 func compileArgs(args []string) (workflowPath, eventPath, format string, err error) {
 	format = "pipeline"
 	filtered := make([]string, 0, len(args))
@@ -2615,54 +2014,6 @@ func compileArgs(args []string) (workflowPath, eventPath, format string, err err
 	}
 	workflowPath, eventPath, err = workflowArgs(filtered)
 	return workflowPath, eventPath, format, err
-}
-
-func executableDigest() (string, error) {
-	_, _, digest, err := executable()
-	return digest, err
-}
-
-func executable() (path string, contents []byte, digest string, err error) {
-	path, err = os.Executable()
-	if err != nil {
-		return "", nil, "", fmt.Errorf("locate compiler executable: %w", err)
-	}
-	readPath := path
-	if runtime.GOOS == "linux" {
-		// /proc/self/exe remains bound to the running inode if the launch path is
-		// replaced. The plugin importer relies on these exact bytes for Linux jobs.
-		readPath = "/proc/self/exe"
-	}
-	file, err := os.Open(readPath)
-	if err != nil {
-		return "", nil, "", fmt.Errorf("open running compiler executable: %w", err)
-	}
-	info, err := file.Stat()
-	if err != nil {
-		_ = file.Close()
-		return "", nil, "", fmt.Errorf("inspect running compiler executable: %w", err)
-	}
-	if !info.Mode().IsRegular() || info.Mode().Perm()&0o111 == 0 || info.Size() <= 0 || info.Size() > runtimeDistributionLimit {
-		_ = file.Close()
-		return "", nil, "", fmt.Errorf("running compiler executable must be an executable regular file between 1 and %d bytes", runtimeDistributionLimit)
-	}
-	contents, err = io.ReadAll(io.LimitReader(file, runtimeDistributionLimit+1))
-	closeErr := file.Close()
-	if err != nil || closeErr != nil {
-		return "", nil, "", fmt.Errorf("read running compiler executable: %w", errors.Join(err, closeErr))
-	}
-	if int64(len(contents)) != info.Size() {
-		return "", nil, "", fmt.Errorf("running compiler executable changed while being read")
-	}
-	platform, err := compiler.ParsePlatform(runtime.GOOS + "/" + runtime.GOARCH)
-	if err != nil {
-		return "", nil, "", fmt.Errorf("validate running compiler executable: %w", err)
-	}
-	if err := validateRuntimeDistributionBinary(platform, contents); err != nil {
-		return "", nil, "", fmt.Errorf("validate running compiler executable: %w", err)
-	}
-	sum := sha256.Sum256(contents)
-	return path, contents, fmt.Sprintf("sha256:%x", sum), nil
 }
 
 func workflowArgs(args []string) (workflowPath, eventPath string, err error) {
